services/govsign/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
liboqs-python>=0.10.0
redis>=5.0.0
sqlalchemy>=2.0.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
cryptography>=42.0.0
